Maintenance - Fix bugs in SyncOWMService - Fix certain models used in GSON conversion - Remove redundant content_main.xml, moving content to activity_main.xml - Tweak preference xml

+++ app/src/main/java/com/kodatos/cumulonimbus/apihelper/models/Main.java
@@ -29,6 +29,6 @@ public class Main {
     public long humidity;
     @SerializedName("temp_kf")
     @Expose
-    public long tempKf;
+    public float tempKf;
 
 }
